Dynamic charts!
The trick we use here is to run the SQL query from within an AJAX call
and control the whole application display from the Javascript side of
things, see web/js/query.js.

That allows fetching the data in the browser and then displaying it as
we want, provided that the display code is written in javascript.

Early output includes HTML table, bar chart, pie chart and donut chart.

For now the query column to use in the charts display is hardcoded to
fit with our query, this will need to be a setup along the way, maybe
some more inputs, maybe a convention trick with the expected names of
the columns in the query output.