Clean up the model, switch back from base36 to integer.
We're still going to use base36 numbers in the URL, but we're going to
manage that in the application. It would be better to use base36 in the
database too so that the URL translates directly to the ID in there, but
that's an extra dependency we'd rather avoid for now.