Improve debian package again, and switch to non-free.
The highcharts source dependency is CC-BY-NC and the NC part means it's
basically non-compliant with debian main DFSG, and as such it must enter
non-free.

It is considered a good enough reason to switch to another charting js
lib, but that is a serious amount of work and let's have 1.x versions in
non-free at the moment.